Rename types.py to scuba_constants.py
